refactor: delete v2ecoli/viz/, migrate remaining callers to visualizations/_helpers
After Tasks 2-8 ported every report's rendering to a Visualization Step,
the legacy v2ecoli/viz/network.py is no longer reachable from any Step.
The two remaining callers (scripts/viz_network.py, reports/workflow_report.py)
are updated to import the same helpers from v2ecoli.visualizations._helpers.
scripts/viz_baseline.py had no v2ecoli.viz import and required no changes.

This completes the atomic refactor — there is no half-state where both
viz APIs are usable.
