--- v2ecoli/viz/__init__.py
@@ -1,17 +1,0 @@
-"""Visualization utilities for v2ecoli composites."""
-
-from v2ecoli.viz.network import (
-    build_graph,
-    render_html,
-    write_outputs,
-    BIO_COLORS,
-    classify,
-)
-
-__all__ = [
-    'build_graph',
-    'render_html',
-    'write_outputs',
-    'BIO_COLORS',
-    'classify',
-]
